// src/types/BLEEnum.ts
export enum BLEEnum {
    SERVICE_UUID = 'ec2e0883-782d-433b-9a0c-6d5df5565410',
    // write characteristic (used to send WiFi credentials)
    CHARACTERISTIC_UUID = 'c2433dd7-137e-4e82-845e-a40f70dc4a8d',
    // notify characteristic (device status notifications)
    NOTIFY_CHARACTERISTIC_UUID = 'c2433dd7-137e-4e82-845e-a40f70dc4a8e'
}